+++ src/main/java/lottogame/domain/machine/LottoDrawingMachine.java
@@ -1,6 +1,8 @@
-package lottogame.domain;
+package lottogame.domain.machine;
 
 import java.util.HashSet;
+import lottogame.domain.number.LottoNumber;
+import lottogame.domain.number.LottoNumbers;
 
 public class LottoDrawingMachine {
