+++ src/main/java/lottogame/domain/number/LottoNumbers.java
@@ -1,4 +1,4 @@
-package lottogame.domain;
+package lottogame.domain.number;
 
 import java.util.ArrayList;
 import java.util.List;
