+++ src/main/java/lottogame/domain/ticket/LottoTicket.java
@@ -1,8 +1,10 @@
-package lottogame.domain;
+package lottogame.domain.ticket;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import lottogame.domain.number.LottoNumber;
+import lottogame.domain.number.LottoNumbers;
 
 public class LottoTicket {
