+++ src/test/java/lottogame/domain/machine/LottoDrawingMachineTest.java
@@ -1,8 +1,9 @@
-package lottogame.domain;
+package lottogame.domain.machine;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lottogame.domain.number.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
