+++ src/test/java/lottogame/domain/number/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lottogame.domain;
+package lottogame.domain.number;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
