+++ src/test/java/lottogame/domain/number/LottoNumbersTest.java
@@ -1,9 +1,11 @@
-package lottogame.domain;
+package lottogame.domain.number;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
+import lottogame.domain.number.LottoNumber;
+import lottogame.domain.number.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
