+++ src/test/java/lottogame/domain/ticket/LottoTicketTest.java
@@ -1,9 +1,10 @@
-package lottogame.domain;
+package lottogame.domain.ticket;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.HashSet;
 import java.util.Set;
+import lottogame.domain.number.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
