+++ src/test/java/lottogame/domain/ticket/LottoTicketsTest.java
@@ -1,7 +1,9 @@
-package lottogame.domain;
+package lottogame.domain.ticket;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lottogame.domain.ticket.LottoTicket;
+import lottogame.domain.ticket.LottoTickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
